lora refactor replay - part 2
- unify LoRa utilities and eliminate multiplex wrapper
- create unified API for batch and serve both to use
- move serve-specific utils
- eliminate dups and cross dependencies
- update imports and tests, test with release/probes
