+++ python/ray/llm/tests/serve/cpu/deployments/llm/multiplex/test_multiplex_utils.py
@@ -3,7 +3,7 @@ from unittest.mock import Mock, call, patch
 
 import pytest
 
-from ray.llm._internal.serve.deployments.llm.multiplex.utils import (
+from ray.llm._internal.common.utils.lora_utils import (
     retry_with_exponential_backoff,
 )
