+++ python/ray/llm/tests/serve/mocks/mock_vllm_engine.py
@@ -4,6 +4,7 @@ import random
 from random import randint
 from typing import AsyncGenerator, Dict, Union
 
+from ray.llm._internal.common.utils.lora_utils import LoraModelLoader
 from ray.llm._internal.serve.configs.openai_api_models import (
     ChatCompletionRequest,
     ChatCompletionResponse,
@@ -18,9 +19,6 @@ from ray.llm._internal.serve.configs.server_models import (
     LLMConfig,
 )
 from ray.llm._internal.serve.deployments.llm.llm_engine import LLMEngine
-from ray.llm._internal.serve.deployments.llm.multiplex.lora_model_loader import (
-    LoraModelLoader,
-)
 
 
 class MockVLLMEngine(LLMEngine):
